pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aml-detection-system"
version = "1.0.0"
description = "Anti-Money Laundering Detection System for Research and Education"
authors = [{name = "AI Research Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "statsmodels>=0.14.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "xgboost>=1.7.0",
    "lightgbm>=4.0.0",
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "tsfresh>=0.20.0",
    "cvxpy>=1.3.0",
    "arch>=6.2.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "beautifulsoup4>=4.12.0",
    "spacy>=3.6.0",
    "networkx>=3.1.0",
    "pyod>=1.1.0",
    "streamlit>=1.25.0",
    "gradio>=3.40.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
    "mlflow>=2.5.0",
    "wandb>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    "pytest>=7.4.0",
    "pre-commit>=3.3.0",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
